python/requirements.txt
ipython==6.4.0
libconf==1.0.1
matplotlib==2.2.2
netifaces==0.10.7
numpy==1.14.3
protobuf==3.5.2
python-daemon==2.1.2
pyzmq==17.0.0
scipy==1.1.0
